Entanglement Pattern
Make the keeping rough track of the set partition of entanglement work. If a gate acts on multiple qubits then it is assumed that it has entangled the blocks. Even if it is an identity gate etc. Better off safe and assume worst case. Measurement is taken care of as well via boolean flag.